Attempt to make it easier to diagnose extension builing problems
