+++ crates/zed/Cargo.toml
@@ -57,6 +57,7 @@ diagnostics.workspace = true
 editor.workspace = true
 env_logger.workspace = true
 extension.workspace = true
+extension_cli.workspace = true
 extension_host.workspace = true
 extensions_ui.workspace = true
 feature_flags.workspace = true
@@ -151,6 +152,7 @@ theme_selector.workspace = true
 time.workspace = true
 title_bar.workspace = true
 toolchain_selector.workspace = true
+tokio.workspace = true
 ui.workspace = true
 ui_input.workspace = true
 ui_prompt.workspace = true
